Improve album root handling
Photo File API was failing in Digikam DBs with more than one AlbumRoot,
code was using the wrong one when building the full path to the file.

Instead of going for the quick fix this improves the handling of album
roots overall, so the ROOT_MAP setting takes precedence, and otherwise
the Digikam DB value is used (AlbumRoots.specificPath).

I include tests for the new root methods in DigikamSQLite, and the test
Digikam DB has been updated with more than one album root.